Move the `stringToPDFString` helper function into the `src/core/string_utils.js` file
Given that this function is only ever used during *parsing* of the PDF document, which happens in the worker-thread, this has always added (a little bit of) dead code in the built `pdf.mjs` file.
