+++ src/core/struct_tree.js
@@ -16,15 +16,14 @@
 import {
   AnnotationPrefix,
   makeArr,
-  stringToPDFString,
   stringToUTF8String,
   warn,
 } from "../shared/util.js";
 import { Dict, isName, Name, Ref, RefSetCache } from "./primitives.js";
+import { stringToAsciiOrUTF16BE, stringToPDFString } from "./string_utils.js";
 import { BaseStream } from "./base_stream.js";
 import { lookupNormalRect } from "./core_utils.js";
 import { NumberTree } from "./name_number_tree.js";
-import { stringToAsciiOrUTF16BE } from "./string_utils.js";
 
 const MAX_DEPTH = 40;
